backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.115.13
h11==0.16.0
idna==3.10
mysql-connector-python==9.3.0
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
